Fix Codex transcript filtering for waits
